Add Alembic; implement is_admin_only sections
__Problem__

We want to limit control access to admin sections. This is kind of the
whole point of the app. :D

But we also don't want to be resetting the database on a whim. Let's get
this thing in prod!

__Solution__

Implement Alembic to track migrations, then add our first migration for
admin sections.

This was kind of difficult to break up since knowing that the migrations
work depended on a migration existing. I just kept at a WIP until I got
the whole thing working (soup to nuts).