Bugfix sections don't lock when created with permissions
 # Problem

When trying to create a section with a permission attached, the API
would bomb out with a database lock error. Documented in #40

 # Solution

We had a nested transaction when trying to add section permissions.
By removing the permission add from the transaction block we allow
the set_permissions function to start its own modifying connection.
Solved!